Rename CryptoStreamGetter to CryptoStreamFactory
