+++ Runtime/ResourceProviders/ICryptoStreamFactory.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e55643bee55f8a54abfb5f982751a0cf
+guid: 4c23770cda40d2d4b8546ab72e788fd5
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
