--- Runtime/ResourceProviders/ICryptoStreamGetter.cs
@@ -1,12 +1,0 @@
-using System.IO;
-using System.Security.Cryptography;
-using UnityEngine.ResourceManagement.ResourceProviders;
-
-namespace Extreal.Integration.Assets.Addressables
-{
-    public interface ICryptoStreamGetter
-    {
-        CryptoStream GetEncryptStream(Stream baseStream, AssetBundleRequestOptions options);
-        CryptoStream GetDecryptStream(Stream baseStream, AssetBundleRequestOptions options);
-    }
-}
